api_files/backend_copy/requirements.txt
fastapi
uvicorn
pillow
numpy
tensorflow  # or torch/onnxruntime, depending on your setup
